refactor: split GroupChatView into focused components
- Extract ChatSidebar, ChatMessageList, ChatInput, MentionPopup, GroupCreateDialog
- Extract useChatWebSocket composable
- Create shared type definitions in types/models.ts
- GroupChatView reduced from ~1268 lines to ~250 lines
